+++ CHANGELOG.md
@@ -17,6 +17,7 @@ All notable changes to the Docker Language Server will be documented in this fil
   - textDocument/documentHighlight
     - support highlighting the short form `depends_on` syntax for services ([#70](https://github.com/docker/docker-language-server/issues/70))
     - support highlighting the long form `depends_on` syntax for services ([#71](https://github.com/docker/docker-language-server/issues/71))
+    - support highlighting referenced networks, volumes, configs, and secrets ([#145](https://github.com/docker/docker-language-server/issues/145))
 
 ### Fixed